templates/hf-af3-nvidia-endpoint/requirements.txt
transformers==4.46.0
accelerate==0.34.2
peft==0.14.0
numpy==1.26.4
Pillow
pydub
soundfile
librosa
openai-whisper
ftfy
jiwer
einops
hydra-core
loguru
matplotlib
pytorchvideo==0.1.5
deepspeed==0.15.4
kaldiio
wandb
opencv-python-headless==4.8.0.76
protobuf==3.20.*
termcolor
sentencepiece
